Add files via upload
Secure Data Storage and in Cloud 
To provide encrypted data storage and data sharing system to improve security.  IT specialists turn to the cloud as the best way to store sensitive data. As we store our conﬁdential documents on cloud, we need some type of encryption to be safe from security breaches